+++ src/main/java/com/app/back/mapper/donation/DonationMapper.java
@@ -3,7 +3,9 @@ package com.app.back.mapper.donation;
 import com.app.back.domain.donation.DonationDTO;
 import com.app.back.domain.donation.DonationVO;
 import com.app.back.domain.donation_record.DonationRecordDTO;
+import com.app.back.domain.inquiry.InquiryDTO;
 import com.app.back.domain.post.Pagination;
+import com.app.back.domain.post.Search;
 import com.app.back.domain.review.ReviewDTO;
 import com.app.back.domain.review.ReviewVO;
 import org.apache.ibatis.annotations.Mapper;
@@ -20,6 +22,8 @@ public interface DonationMapper {
     public Optional<DonationDTO> selectById(Long id);
     // 전체 조회
     public List<DonationDTO> selectAll(@Param("pagination") Pagination pagination);
+
+    public List<DonationDTO> selectFilterAll(@Param("pagination") Pagination pagination);
     // 전체 개수
     public int selectCount();
     // 수정
