+++ src/main/resources/mapper/donationMapper.xml
@@ -6,13 +6,28 @@
     </insert>
 
     <select id="selectAll" resultType="donationDTO">
-        select d.id, d.goal_point, d.donation_s_date, d.donation_e_date, p.post_title, p.post_content, p.member_id, p.created_date, p.updated_date, m.member_nickname
+        select d.id, d.goal_point, d.donation_s_date, d.donation_e_date, p.post_title, p.post_content, p.post_view_count, p.member_id, p.created_date, p.updated_date, m.member_nickname
         from tbl_donation d join tbl_post p
             on d.id = p.id
         join tbl_member m
             on p.member_id = m.id
-        order by created_date desc, r.id desc
+        order by created_date desc, d.id desc
         limit #{pagination.startRow}, #{pagination.rowCount};
+    </select>
+
+    <select id="selectFilterAll" resultType="donationDTO">
+        select d.id, d.goal_point, d.donation_s_date, d.donation_e_date, p.post_title, p.post_content, p.post_view_count, p.member_id, p.created_date, p.updated_date, m.member_nickname
+        from tbl_donation d join tbl_post p
+                                 on d.id = p.id
+                            join tbl_member m
+                                 on p.member_id = m.id
+        order by
+        <choose>
+            <when test="pagination.order = '조회순'.toString()">p.post_view_count</when>
+            <otherwise>d.donation_e_date</otherwise>
+        </choose>
+        desc
+            limit #{pagination.startRow}, #{pagination.rowCount};
     </select>
 
     <select id="selectById" resultType="donationDTO">
