+++ src/main/resources/templates/donation/donation-update.html
@@ -236,5 +236,8 @@
             </div>
         </div>
     </body>
+    <script th:inline="javascript">
+        const donation = [[${donation}]];
+    </script>
     <script src="/js/donation/donation-update.js"></script>
 </html>
